refactor(SidebarE): Use partials for styled-components
