+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/Sidebar.tsx
@@ -1,6 +1,4 @@
 import React from 'react'
-import styled from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 import { Transition } from 'react-transition-group'
 
 import { SidebarHandle } from './SidebarHandle'
@@ -10,6 +8,10 @@ import { ComponentWithClass } from '../../../common/ComponentWithClass'
 import { SidebarContext, SidebarProvider } from './context'
 import { NotificationsProps } from '../NotificationPanel'
 import { TRANSITION_STYLES, TRANSITION_TIMEOUT } from './constants'
+import { StyledSidebar } from './partials/StyledSidebar'
+import { StyledHead } from './partials/StyledHead'
+import { StyledHeadIcon } from './partials/StyledHeadIcon'
+import { StyledHeadTitle } from './partials/StyledHeadTitle'
 
 export interface SidebarEProps extends ComponentWithClass {
   icon?: React.ReactNode
@@ -18,60 +20,6 @@ export interface SidebarEProps extends ComponentWithClass {
   hasUnreadNotification?: boolean
   notifications?: React.ReactElement<NotificationsProps>
 }
-
-interface StyledSidebarProps {
-  isOpen?: boolean
-}
-
-const { color, spacing, fontSize, zIndex } = selectors
-
-const StyledSidebar = styled.aside<StyledSidebarProps>`
-  display: inline-flex;
-  flex-direction: column;
-  position: relative;
-  z-index: ${zIndex('sidebar', 0)};
-  width: ${({ isOpen }) => (isOpen ? '18rem' : '3.75rem')};
-  height: 100vh;
-  background-color: ${color('neutral', '700')};
-  color: ${color('neutral', 'white')};
-  transition: ${({ isOpen }) =>
-    isOpen
-      ? '200ms width cubic-bezier(0.34, 1.56, 0.64, 1)'
-      : '200ms width cubic-bezier(0.34, 1, 0.64, 1)'};
-
-  > a:hover {
-    text-decoration: none;
-  }
-`
-
-const StyledHead = styled.div`
-  display: flex;
-  align-items: center;
-  padding: ${spacing('8')} ${spacing('6')};
-  background-color: ${color('neutral', '800')};
-`
-
-const StyledIcon = styled.div`
-  display: inline-flex;
-  align-items: center;
-  border-radius: 4px;
-  background-color: ${color('neutral', '500')};
-  padding: 0.55rem;
-  margin-right: ${spacing('7')};
-
-  svg {
-    width: 18px;
-    height: 18px;
-  }
-`
-
-const StyledTitle = styled.div`
-  font-weight: 600;
-  font-size: ${fontSize('l')};
-  white-space: nowrap;
-  opacity: 1;
-  transition: opacity 150ms linear;
-`
 
 export const SidebarE: React.FC<SidebarEProps> = ({
   icon,
@@ -97,16 +45,16 @@ export const SidebarE: React.FC<SidebarEProps> = ({
             </Transition>
             {title && (
               <StyledHead data-testid="sidebar-head">
-                {icon && <StyledIcon>{icon}</StyledIcon>}
+                {icon && <StyledHeadIcon>{icon}</StyledHeadIcon>}
                 <Transition
                   in={isOpen}
                   timeout={TRANSITION_TIMEOUT}
                   unmountOnExit
                 >
                   {(state) => (
-                    <StyledTitle style={{ ...TRANSITION_STYLES[state] }}>
+                    <StyledHeadTitle style={{ ...TRANSITION_STYLES[state] }}>
                       {title}
-                    </StyledTitle>
+                    </StyledHeadTitle>
                   )}
                 </Transition>
               </StyledHead>
