+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/SidebarNavItem.tsx
@@ -1,6 +1,4 @@
 import React, { useState, useContext } from 'react'
-import styled, { css } from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 import { Transition } from 'react-transition-group'
 
 import { SidebarSubNav } from './SidebarSubNav'
@@ -8,6 +6,9 @@ import { Nav, NavItem } from '../../../common/Nav'
 import { SidebarContext } from './context'
 import { Tooltip } from '../../Tooltip'
 import { TRANSITION_STYLES, TRANSITION_TIMEOUT } from './constants'
+import { StyledNavItem } from './partials/StyledNavItem'
+import { StyledNavItemIcon } from './partials/StyledNavItemIcon'
+import { StyledNavItemText } from './partials/StyledNavItemText'
 
 export interface SidebarNavItemEProps
   extends NavItem,
@@ -15,93 +16,6 @@ export interface SidebarNavItemEProps
   icon?: React.ReactNode
   onClick?: (e: React.MouseEvent<HTMLElement>) => void
 }
-
-interface StyledSidebarNavItemProps {
-  isActive?: boolean
-}
-
-interface StyledIconProps {
-  isOpen?: boolean
-}
-
-interface StyledTextProps {
-  isOpen?: boolean
-}
-
-const { color, spacing, fontSize } = selectors
-
-export const StyledSidebarNavItem = styled.div<StyledSidebarNavItemProps>`
-  display: flex;
-  align-items: center;
-  margin: ${spacing('4')} 0;
-  border-radius: 2px;
-
-  > *:first-child {
-    display: flex;
-    align-items: center;
-    width: 100%;
-  }
-
-  ${({ isActive }) =>
-    isActive &&
-    css`
-      background-color: ${color('action', '500')};
-
-      &:hover {
-        background-color: ${color('action', '500')};
-      }
-    `}
-
-  &:hover {
-    background-color: ${color('neutral', '400')};
-
-    > * {
-      text-decoration: none;
-    }
-  }
-`
-
-const StyledIcon = styled.div<StyledIconProps>`
-  display: inline-flex;
-  align-items: center;
-  width: 100%;
-  padding: 0.55rem;
-
-  svg {
-    width: 18px;
-    height: 18px;
-    color: ${color('neutral', '100')};
-
-    ${StyledSidebarNavItem}:hover & {
-      color: ${color('neutral', 'white')};
-    }
-  }
-
-  ${({ isOpen }) =>
-    isOpen &&
-    css`
-      width: auto;
-    `}
-`
-
-const StyledText = styled.div<StyledTextProps>`
-  display: inline-block;
-  color: ${color('neutral', '100')};
-  font-size: ${fontSize('m')};
-  margin-left: ${spacing('2')};
-  opacity: 1;
-  transition: opacity 150ms linear;
-
-  ${StyledSidebarNavItem}:hover & {
-    color: ${color('neutral', 'white')};
-  }
-
-  ${({ isOpen }) =>
-    !isOpen &&
-    css`
-      display: none;
-    `}
-`
 
 export const SidebarNavItemE: React.FC<SidebarNavItemEProps> = ({
   isActive,
@@ -118,16 +32,16 @@ export const SidebarNavItemE: React.FC<SidebarNavItemEProps> = ({
     ...link.props,
     children: (
       <>
-        {icon && <StyledIcon isOpen={isOpen}>{icon}</StyledIcon>}
+        {icon && <StyledNavItemIcon isOpen={isOpen}>{icon}</StyledNavItemIcon>}
         <Transition in={isOpen} timeout={TRANSITION_TIMEOUT} unmountOnExit>
           {(state) => (
-            <StyledText
+            <StyledNavItemText
               style={{ ...TRANSITION_STYLES[state] }}
               isOpen={isOpen}
               data-testid="sidebar-nav-item-text"
             >
               {linkElement.props.children}
-            </StyledText>
+            </StyledNavItemText>
           )}
         </Transition>
         {!isOpen && hasMouseOver && (
@@ -143,14 +57,14 @@ export const SidebarNavItemE: React.FC<SidebarNavItemEProps> = ({
   })
 
   return (
-    <StyledSidebarNavItem
+    <StyledNavItem
       isActive={isActive}
       onMouseEnter={(_) => setHasMouseOver(true)}
       onMouseLeave={(_) => setHasMouseOver(false)}
     >
       {item}
       {isOpen && children && <SidebarSubNav>{children}</SidebarSubNav>}
-    </StyledSidebarNavItem>
+    </StyledNavItem>
   )
 }
