+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/SidebarSubNav.tsx
@@ -1,76 +1,21 @@
 import React from 'react'
-import styled from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 import { IconMoreVert } from '@royalnavy/icon-library'
 
 import { SHEET_PLACEMENT } from '../Sheet/constants'
-import { Sheet } from '../Sheet/Sheet'
-import { SheetButton } from '../Sheet/SheetButton'
 import { Nav } from '../../../common/Nav'
 import { SidebarNavItemEProps } from './SidebarNavItem'
+import { StyledSubNavSheet } from './partials/StyledSubNavSheet'
+import { StyledSubNavSheetButton } from './partials/StyledSubNavSheetButton'
 
 const SHEET_WIDTH = 200
-
-const { color, spacing } = selectors
-
-const StyledSheet = styled(Sheet)`
-  display: flex;
-
-  > div > div {
-    margin-left: ${spacing('3')};
-  }
-
-  ol {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-  }
-
-  a {
-    padding: ${spacing('2')};
-  }
-
-  a,
-  a:hover {
-    color: ${color('neutral', 'white')};
-    text-decoration: none;
-  }
-`
-
-const StyledSheetButton = styled(SheetButton)`
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  height: 18px;
-  width: 18px;
-  padding: 0;
-  margin-right: ${spacing('4')};
-  border-radius: 2px;
-  border: none;
-  background-color: transparent;
-  overflow: hidden;
-
-  &:hover {
-    background-color: ${color('neutral', '400')};
-    cursor: pointer;
-
-    svg {
-      color: ${color('neutral', 'white')};
-    }
-  }
-
-  svg {
-    color: ${color('neutral', '100')};
-  }
-`
 
 export const SidebarSubNav: React.FC<Nav<SidebarNavItemEProps>> = ({
   children,
 }) => {
   return (
-    <StyledSheet
+    <StyledSubNavSheet
       button={(
-        <StyledSheetButton
+        <StyledSubNavSheetButton
           aria-label="Expand sub-menu"
           data-testid="sub-menu-expand-button"
           icon={<IconMoreVert />}
@@ -82,7 +27,7 @@ export const SidebarSubNav: React.FC<Nav<SidebarNavItemEProps>> = ({
       data-testid="sidebar-sub-nav"
     >
       <ol>{children}</ol>
-    </StyledSheet>
+    </StyledSubNavSheet>
   )
 }
