+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/SidebarUser.tsx
@@ -1,18 +1,18 @@
 import React, { useContext } from 'react'
-import styled from 'styled-components'
-import { selectors } from '@royalnavy/design-tokens'
 import { IconPerson, IconExitToApp } from '@royalnavy/icon-library'
 import { Transition } from 'react-transition-group'
 
-import { Avatar } from '../../Avatar'
 import { ComponentWithClass } from '../../../common/ComponentWithClass'
 import { LinkTypes } from '../../../common/Link'
 import { SidebarContext } from './context'
 import { SidebarUserItemE } from './SidebarUserItem'
-import { Sheet } from '../Sheet/Sheet'
-import { SheetButton } from '../Sheet/SheetButton'
 import { SHEET_PLACEMENT } from '../Sheet/constants'
 import { TRANSITION_STYLES, TRANSITION_TIMEOUT } from './constants'
+import { StyledUserAvatar } from './partials/StyledUserAvatar'
+import { StyledUserSheet } from './partials/StyledUserSheet'
+import { StyledUserSheetButton } from './partials/StyledUserSheetButton'
+import { StyledUser } from './partials/StyledUser'
+import { StyledUserText } from './partials/StyledUserText'
 
 export interface SidebarUserEProps extends ComponentWithClass {
   children?: never
@@ -26,96 +26,19 @@ type SidebarAvatarWithItemsProps = Omit<SidebarUserEProps, 'link'>
 
 const SHEET_WIDTH = 106
 
-const { spacing, color, fontSize } = selectors
-
-const StyledSidebarUser = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: ${spacing('10')} ${spacing('8')};
-  background-color: ${color('neutral', '500')};
-`
-
-const StyledSidebarUserText = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: calc(100% - 1rem);
-  padding: 0 0 0 ${spacing('6')};
-  white-space: nowrap;
-  opacity: 1;
-  transition: opacity 150ms linear;
-
-  > div {
-    display: inline-flex;
-    justify-content: center;
-    flex-direction: column;
-
-    a:hover {
-      text-decoration: none;
-    }
-
-    > span {
-      color: ${color('neutral', 'white')};
-      font-size: ${fontSize('m')};
-    }
-
-    a span {
-      margin-top: ${spacing('1')};
-      color: ${color('neutral', '300')};
-      font-size: ${fontSize('s')};
-    }
-  }
-
-  svg {
-    width: 1.65rem;
-    height: 1.65rem;
-    color: ${color('neutral', 'white')};
-    background-color: ${color('neutral', '400')};
-    border-radius: 2px;
-    padding: ${spacing('2')};
-  }
-`
-
-const StyledAvatar = styled(Avatar)`
-  &:hover {
-    text-decoration: none;
-  }
-`
-
-const StyledSheet = styled(Sheet)`
-  > div > div {
-    margin-left: 1px;
-  }
-
-  ol {
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-  }
-
-  a,
-  a:hover {
-    color: ${color('neutral', 'white')};
-    text-decoration: none;
-  }
-`
-
-const StyledSheetButton = styled(SheetButton)`
-  padding: 0;
-`
-
 const SidebarAvatarWithItems: React.FC<SidebarAvatarWithItemsProps> = ({
   initials,
   userLink,
   exitLink,
 }) => (
-  <StyledSheet
+  <StyledUserSheet
     button={(
-      <StyledSheetButton
+      <StyledUserSheetButton
         aria-label="Show user options"
         data-testid="user-button"
-        icon={<StyledAvatar initials={initials} data-testid="sidebar-avatar" />}
+        icon={
+          <StyledUserAvatar initials={initials} data-testid="sidebar-avatar" />
+        }
       />
     )}
     placement={SHEET_PLACEMENT.RIGHT_BOTTOM}
@@ -127,7 +50,7 @@ const SidebarAvatarWithItems: React.FC<SidebarAvatarWithItemsProps> = ({
         <SidebarUserItemE icon={<IconExitToApp />} link={exitLink} />
       )}
     </ol>
-  </StyledSheet>
+  </StyledUserSheet>
 )
 
 export const SidebarUserE: React.FC<SidebarUserEProps> = ({
@@ -140,24 +63,22 @@ export const SidebarUserE: React.FC<SidebarUserEProps> = ({
 
   if (!isOpen) {
     return (
-      <StyledSidebarUser data-testid="sidebar-user-closed-children">
+      <StyledUser data-testid="sidebar-user-closed-children">
         <SidebarAvatarWithItems
           initials={initials}
           userLink={userLink}
           exitLink={exitLink}
         />
-      </StyledSidebarUser>
+      </StyledUser>
     )
   }
 
   return (
-    <StyledSidebarUser
-      data-testid={`sidebar-user-${isOpen ? 'open' : 'closed'}`}
-    >
-      <StyledAvatar initials={initials} />
+    <StyledUser data-testid={`sidebar-user-${isOpen ? 'open' : 'closed'}`}>
+      <StyledUserAvatar initials={initials} />
       <Transition in={isOpen} timeout={TRANSITION_TIMEOUT} unmountOnExit>
         {(state) => (
-          <StyledSidebarUserText style={{ ...TRANSITION_STYLES[state] }}>
+          <StyledUserText style={{ ...TRANSITION_STYLES[state] }}>
             <div>
               <span>{name}</span>
               {userLink &&
@@ -172,10 +93,10 @@ export const SidebarUserE: React.FC<SidebarUserEProps> = ({
                 children: <IconExitToApp />,
                 'data-testid': 'sidebar-exit-link',
               })}
-          </StyledSidebarUserText>
+          </StyledUserText>
         )}
       </Transition>
-    </StyledSidebarUser>
+    </StyledUser>
   )
 }
