+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/SidebarWrapper.tsx
@@ -1,22 +1,13 @@
 import React from 'react'
-import styled from 'styled-components'
 
 import { ComponentWithClass } from '../../../common/ComponentWithClass'
-
-const StyledSidebarWrapper = styled.div`
-  display: flex;
-  flex-direction: row;
-`
+import { StyledWrapper } from './partials/StyledWrapper'
 
 export const SidebarWrapperE: React.FC<ComponentWithClass> = ({
   className,
   children,
 }) => {
-  return (
-    <StyledSidebarWrapper className={className}>
-      {children}
-    </StyledSidebarWrapper>
-  )
+  return <StyledWrapper className={className}>{children}</StyledWrapper>
 }
 
 SidebarWrapperE.displayName = 'SidebarWrapper'
