+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledHead.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing, color } = selectors
+
+export const StyledHead = styled.div`
+  display: flex;
+  align-items: center;
+  padding: ${spacing('8')} ${spacing('6')};
+  background-color: ${color('neutral', '800')};
+`
