+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledNotificationsSheet.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+import { Sheet } from '../../Sheet/Sheet'
+
+export const StyledNotificationsSheet = styled(Sheet)`
+  display: flex;
+  justify-content: center;
+`
