+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledUserAvatar.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+import { Avatar } from '../../../Avatar'
+
+export const StyledUserAvatar = styled(Avatar)`
+  &:hover {
+    text-decoration: none;
+  }
+`
