+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledUserItem.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing } = selectors
+
+export const StyledUserItem = styled.li`
+  display: flex;
+  align-items: center;
+  margin: ${spacing('4')};
+`
