+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledUserSheetButton.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+import { SheetButton } from '../../Sheet/SheetButton'
+
+export const StyledUserSheetButton = styled(SheetButton)`
+  padding: 0;
+`
