+++ packages/react-component-library/src/components/TopLevelNavigation/SidebarE/partials/StyledWrapper.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const StyledWrapper = styled.div`
+  display: flex;
+  flex-direction: row;
+`
